Changed mode to 0 on create in Linux variant to get exclusive access
Split Linux and BSD version to make code for each more readable.  Hopefully haven't broken anything by doing so.

On linux, opening the lockfile with --------- creates the file, and the opening process can RW to it, but everyone else gets a permissions error. It is effectively the same as having a BSD O_EXLOCK option